checkpoint

+++ src/mongo/db/pipeline/expression.cpp
@@ -1830,7 +1830,8 @@ namespace mongo {
         const intrusive_ptr<ExpressionFieldPath> &pTheFieldPath, CmpOp cmpOp,
         const intrusive_ptr<const Value> &pValue):
         pFieldPath(pTheFieldPath),
-        pRange(new Range(cmpOp, pValue)) {
+        pRange(new Range(cmpOp, pValue)) {
+
     }
 
     void ExpressionFieldRange::intersect(
@@ -2065,7 +2066,7 @@ namespace mongo {
         ExpressionNary::addOperand(pExpression);
     }
 
-    int ExpressionIsoDate::checkIntRange(const char *pName, long value) const {
+    int ExpressionIsoDate::checkIntRange(const char *pName, long long value) const {
         uassert(16027, str::stream() << getOpName() <<
                 ":  \"" << pName <<
                 "\" value outside of range INT_MIN to INT_MAX",
